+++ src/components/Switch/index.ts
@@ -1,4 +1,4 @@
-// Switch from copy-ui @ 2025-12-07.
+// Switch from copy-ui @ 2025-12-13.
 
 export type { SwitchProps } from "./switch";
 export { Switch } from "./switch";