+++ src/components/Switch/CHANGELOG.md
@@ -9,3 +9,4 @@
   gap values. 
 - 2025-11-03: Use JSS & resolve-style for styles.
 - 2025-12-07: Refactor code and remove deprecated dependencies.
+- 2025-12-13: Update to support the new color scheme.